Implemented Tag NfcV check
Implemented a way to see if the nfc tag is NfcV (ISO 15693) or not.

This is becouse we want to reverse what we read if its NfcV to match it with how IOS handles it.

We dont want to reverse it if its NfcA for an example becouse android and IOS handles the reading the same.